refactor(gsd): extract git checkout/stash recovery from auto-worktree into worktree-git-recovery
checkoutBranchWithStashGuard, removeMergeStateFiles, cleanupConflictState,
and the stash helpers (popStashByRef, stashRefFromError,
stashAlreadyExistsFilesFromError, hasConflictMarkers,
gsdJsonlFilesWithConflictMarkers) were embedded in the 2,600-line
auto-worktree orchestration module — the recurring-bug hot spot for stash
EEXIST collisions, leaked MERGE_HEAD/SQUASH_MSG state, and conflicted
stash pops (7 of the last 30 worktree commits). They now live behind
their own module so recovery fixes land in one place and can be tested
against scripted git states. Code moved verbatim;
checkoutBranchWithStashGuard stays re-exported from auto-worktree for
existing callers. The State Reconciliation drift repair keeps its own
merge-state primitive by design (drift repairs own raw primitives).
